add teleport

+++ VSPlugin/Gamepad/GamepadState.cs
@@ -4,11 +4,11 @@
     {
         public readonly InputX A, B, X, Y, Start, Back, L1, L3, R1, R3, DPadUp, DPadDown, DPadLeft, DPadRight;
         public readonly float L2, R2;
-        public readonly double LeftThumbX, LeftThumbY, RightThumbX, RightThumbY;
+        public readonly float LeftThumbX, LeftThumbY, RightThumbX, RightThumbY;
 
         public GamepadState(InputX A, InputX B, InputX X, InputX Y, InputX Start, InputX Back, InputX L1, float L2,
         InputX L3, InputX R1, float R2, InputX R3, InputX DPadUp, InputX DPadDown, InputX DPadLeft, InputX DPadRight,
-        double LeftThumbX, double LeftThumbY, double RightThumbX, double RightThumbY)
+        float LeftThumbX, float LeftThumbY, float RightThumbX, float RightThumbY)
         {
             this.A = A; this.B = B; this.X = X; this.Y = Y;
             this.Start = Start; this.Back = Back;
